device_manager/device_manager.py
# device_manager.py

class DeviceManager:
    def __init__(self):
        self.devices = {}

    def add_device(self, device_id, device_type, config):
        """Instantiate a device object and store it in the devices dictionary."""
        if device_type == 'radar':
            self.devices[device_id] = RadarSensor(config)
        elif device_type == 'turbidity':
            self.devices[device_id] = TurbiditySensor(config)
        elif device_type == 'ph':
            self.devices[device_id] = PHSensor(config)
        # Add more elif blocks for other device types
        else:
            raise ValueError(f"Unknown device type: {device_type}")

    def get_device(self, device_id):
        """Retrieve a device object by its ID."""
        return self.devices.get(device_id)

    def read_device(self, device_id):
        """Read data from a specific device."""
        device = self.get_device(device_id)
        if device:
            return device.read_data()
        else:
            raise ValueError(f"Device with ID {device_id} not found")

    def write_device(self, device_id, data):
        """Write data to a specific device."""
        device = self.get_device(device_id)
        if device:
            device.write_data(data)
        else:
            raise ValueError(f"Device with ID {device_id} not found")

    def update_device_config(self, device_id, config):
        """Update the configuration of a specific device."""
        device = self.get_device(device_id)
        if device:
            device.update_configuration(config)
        else:
            raise ValueError(f"Device with ID {device_id} not found")

# Example usage
if __name__ == "__main__":
    manager = DeviceManager()
    manager.add_device('radar1', 'radar', {'model': 'RadarX', 'serial_number': '12345'})
    manager.add_device('turbidity1', 'turbidity', {'model': 'TurbidityX', 'serial_number': '67890'})
    
    radar_data = manager.read_device('radar1')
    turbidity_data = manager.read_device('turbidity1')
    
    print(f"Radar Data: {radar_data}")
    print(f"Turbidity Data: {turbidity_data}")

    # Write data or update configuration as needed
    manager.write_device('radar1', {'operation_mode': 'automatic'})
    manager.update_device_config('turbidity1', {'calibration_value': 3.14})



""" Here's how the structure of the device classes might look, inheriting from a common DeviceInterface. 
Each specific device class (e.g., RadarSensor, TurbiditySensor, PHSensor) would implement the methods defined in the DeviceInterface."""
# device_interface.py

from abc import ABC, abstractmethod

class DeviceInterface(ABC):
    @abstractmethod
    def read_data(self):
        """Read data from the device."""
        pass

    @abstractmethod
    def write_data(self, data):
        """Write data to the device."""
        pass

    @abstractmethod
    def update_configuration(self, config):
        """Update the device configuration."""
        pass

"""Then, for each specific device, we would have a separate module. For instance, the radar_sensor.py module might look like this:"""
# radar_sensor.py

from device_interface import DeviceInterface

class RadarSensor(DeviceInterface):
    def __init__(self, config):
        # Initialize the radar sensor with the given configuration
        self.config = config

    def read_data(self):
        # Implement the reading mechanism specific to the radar sensor
        return {"distance": 123.45}

    def write_data(self, data):
        # Implement the writing mechanism specific to the radar sensor
        pass

    def update_configuration(self, config):
        # Implement the update configuration mechanism specific to the radar sensor
        self.config.update(config)
"""Similarly, we would define turbidity_sensor.py and ph_sensor.py modules for other types of sensors.

Finally, you would import and use these device classes in your device_manager.py:"""
# device_manager.py

from device_manager.radar_sensor import RadarSensor
from device_manager.turbidity_sensor import TurbiditySensor
from device_manager.ph_sensor import PHSensor
# ...import other device classes as needed

# The rest of the DeviceManager code remains the same as the previous example

