tests/test_device_manager.py
#This script uses a mock device class to simulate interactions with real devices. 
#It defines a series of tests to check if the DeviceManager can add devices, read data, write data, and handle errors as expected.
#The MockDeviceWithState class is an extension of the MockDevice that holds state, allowing us to test the device manager's ability to interact with devices whose state can change.

import unittest
from device_manager import DeviceManager, DeviceInterface

# Mock classes for testing
class MockDevice(DeviceInterface):
    def read_data(self):
        return "mock data"

    def write_data(self, data):
        pass

    def update_configuration(self, config):
        pass

class TestDeviceManager(unittest.TestCase):

    def setUp(self):
        # Set up the DeviceManager and add a few mock devices for testing
        self.dev_manager = DeviceManager()
        self.dev_manager.add_device('radar', MockDevice())
        self.dev_manager.add_device('ph', MockDevice())

    def test_add_device(self):
        # Test that devices are correctly added to the manager
        self.assertIn('radar', self.dev_manager.devices)
        self.assertIn('ph', self.dev_manager.devices)
        self.assertIsInstance(self.dev_manager.get_device('radar'), MockDevice)

    def test_read_data(self):
        # Test reading data from a device
        data = self.dev_manager.get_device('radar').read_data()
        self.assertEqual(data, "mock data")

    def test_write_data(self):
        # Test writing data to a device
        self.dev_manager.get_device('ph').write_data("new data")
        # Since MockDevice doesn't actually store data, this is just to ensure no errors

    def test_update_configuration(self):
        # Test updating the configuration of a device
        self.dev_manager.get_device('ph').update_configuration({'new': 'config'})
        # Similar to write_data, this is just to ensure the method runs without errors

    def test_device_not_found(self):
        # Test that the appropriate error is raised when a non-existent device is accessed
        with self.assertRaises(KeyError):
            self.dev_manager.get_device('non_existent')

    # Additional mock device for testing concurrency and duplicate device types
class MockDeviceWithState(MockDevice):
    def __init__(self):
        self.data = "initial data"

    def read_data(self):
        return self.data

    def write_data(self, data):
        self.data = data

# Continuing from the previous TestDeviceManager class...

    def test_concurrent_device_access(self):
        # Test that the device manager can handle concurrent access to devices
        radar_device = self.dev_manager.get_device('radar')
        radar_device.write_data("concurrent access test")
        data = radar_device.read_data()
        self.assertEqual(data, "concurrent access test")

    def test_duplicate_device_type(self):
        # Test managing multiple devices of the same type
        self.dev_manager.add_device('radar2', MockDeviceWithState())
        self.dev_manager.get_device('radar2').write_data("radar 2 data")
        data = self.dev_manager.get_device('radar2').read_data()
        self.assertEqual(data, "radar 2 data")
        # Ensure that data from radar2 does not interfere with the original radar device
        original_radar_data = self.dev_manager.get_device('radar').read_data()
        self.assertEqual(original_radar_data, "mock data")

    def test_error_handling(self):
        # Test the device manager's error handling during a device failure
        def faulty_read_data():
            raise Exception("Device communication failure")

        # Replace the read_data method of a device with a faulty one
        faulty_device = self.dev_manager.get_device('ph')
        faulty_device.read_data = faulty_read_data

        with self.assertRaises(Exception) as context:
            faulty_device.read_data()
        self.assertTrue('Device communication failure' in str(context.exception))

    def test_device_removal(self):
        # Test removing a device
        self.dev_manager.remove_device('radar')
        self.assertNotIn('radar', self.dev_manager.devices)

    def test_device_update(self):
        # Test updating the configuration of a device that changes its behavior
        radar_device = self.dev_manager.get_device('radar')
        radar_device.update_configuration({'data': 'updated data'})
        data = radar_device.read_data()
        self.assertEqual(data, 'updated data')

# ... you can continue adding more tests as needed.

if __name__ == '__main__':
    unittest.main()
